Refactor error pages
